Revert "refactor: move activate job record into the proper class"
This reverts commit 032c5589
